+++ store/user-service/src/main/java/by/innowise/userservice/exception/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package by.innowise.userservice.exception.runtime;
+package by.innowise.userservice.exception;
 
 public class EntityNotFoundException extends RuntimeException {
