+++ store/user-service/src/main/java/by/innowise/userservice/repository/CardInfoRepository.java
@@ -1,10 +1,9 @@
 package by.innowise.userservice.repository;
 
-import by.innowise.userservice.exception.runtime.EntityNotFoundException;
+import by.innowise.userservice.exception.EntityNotFoundException;
 import by.innowise.userservice.model.entity.CardInfo;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
